requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==4.1.5
django-cacheops==6.1
django-pg-bulk-update==3.7.0
django-pg-returning==2.0.0
funcy==1.17
psycopg2==2.9.5
redis==4.4.2
six==1.16.0
sqlparse==0.4.3
